Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==2.2.5
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-client==0.12.0
flask_cors==4.0.0

# ✅ Fix: Upgrade to latest Prometheus Exporter (>=0.22.0)
prometheus-flask-exporter>=0.22.0

# OpenTelemetry (Aligned Versions)
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-semantic-conventions==0.38b0